added initial comments

+++ 2homework/1.1/Point.cpp
@@ -4,15 +4,17 @@
 
 using namespace std;
 
-
+// Constructors:
 // Default Constructor (No parameters)
+// Assignes 0 to X and Y as the default value
 Point::Point() : x(0), y(0) 
 {
     cout << "in point default constructor" << endl;
 }
 
 // Constructor (With parameters)
-Point::Point(float x, float y) : x(x), y(y) 
+// Assigmes the argument values to X and Y
+Point::Point(double x, double y) : x(x), y(y) 
 {
     cout << "in point two parameter constructor" << endl;
 }
@@ -27,35 +29,38 @@ Point::Point(Point &p)
 }
 
 
-// Setters
-void Point::setX(float n)
+// Setters:
+// X Setter
+void Point::setX(double n)
 {
     this->x = n;
 }
 
-
-void Point::setY(float n)
+// Y Setter
+void Point::setY(double n)
 {
     this->y = n;
 }
 
 
-// Getters
-const float& Point::getX() const
+// Getters:
+// X Getter
+const double& Point::getX() const
 {
     return this->x;
 }
 
-const float& Point::getY() const
+// Y Getter
+const double& Point::getY() const
 {
     return this->y;
 }
 
-
-float Point::distance(Point &p) const
+// 
+double Point::distance(Point &p) const
 {
-    float dx = this->x - p.getX();
-    float dy = this->y - p.getY();
+    double dx = this->x - p.getX();
+    double dy = this->y - p.getY();
 
     return sqrt((dx * dx) + (dy * dy));
 }
